Have managed to split NpgsqlDataReader into abstract API + implementation (passing all tests)
